Edited persistence layer. added version property for each entity for later versioning of database tables. Fixed loading surgeryunit with available timeslots.

+++ src/main/java/de/bbajor/pvs/ivomplan/model/IvomDiagnosis.java
@@ -10,7 +10,7 @@ import lombok.experimental.Accessors;
 @EqualsAndHashCode(callSuper = true)
 @Entity
 @Accessors(chain = true)
-public class IvomDiagnosis extends BasicEntity<Integer> {
+public class IvomDiagnosis extends BasicEntity<Long> {
 
     private String name;
     private String icdCode;
